Make a transaction-std-test-utils module, to help code reuse
This will be useful for testing the signed contingent inputs
